q2/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai==0.3.2
python-dotenv==1.0.1
pillow==10.2.0
pydantic==2.6.1 